Make all prefereces panes scrollable
Since the preferences window can grow too tall for some screens,
ensure that all of the notebook tabs are wrapped in a ScrolledWindow
so that the dialog can grow as tall or short as needed (above an
arbitrarily-chosen minimum height of 400px).

Disable overlay scrolling so that if the current tab IS scrollable,
it's very obvious that there are more preferences than what's visible.