pdf_multi: Work around image-extraction bug
Some PDFs with embedded images exhibit what I believe to be
(and have reported as) a bug in PyMuPDF[1], where a call to
fitz.image_profile() for a valid embedded image xref returns
None instead of the documented empty dict ({}). This change
works around that by eliminating the assumption that `.get()`
can always be called on the return value.

[1]: https://github.com/pymupdf/PyMuPDF/issues/2501