Replace translation monkeypatching with import
Monkeypatching builtins to "install" the _() function is frowned upon,
and unsupported by mypy[1]. It's also unnecessary, as it's perfectly
workable to create a _() function definition and import it where needed.
This commit adds `from mcomix.i18n import _` to the imports of all
modules that use translations, replacing the monkeypatched global.
[1]: https://github.com/python/mypy/issues/8727